Update Node Binary Installation Guide (#874)
* Update node binary installation guide for new CLI version (0.3.x)

This commit will deprecate old CLI. Install guide with binaries and ansible will be marked as legacy. A new guide page will be created to ask users to install binaries with an installation script directly. The new guide will make node binary installation and update easier for users.

* Apply suggestions from code review

Co-authored-by: Danny Salman <salman.danny03@hotmail.com>

* Apply suggestions from code review

Co-authored-by: Danny Salman <salman.danny03@hotmail.com>

* Mark v0.3.0 installation doc as Beta before releasing to mainnet

We will deploy v0.3.0 to Mumbai testnet and wait for a few weeks before continue to mainnet. During this time, the installation doc should keep the old version and show new guide as beta. Once we are ready for mainnet, we can revert this commit.

* Rename 'Migration' section title to 'Upgrade'

* Restore keystore to original location

* Restore keystore to original location

* Minor fix

* Update v0.3.0 docs according to new changes in CLI and config.

* Update install script for v0.3.0

* Minor fixes

* Deprecate old 0.2.x pages

* Apply suggestions from review

* Update full-node.md

* Apply suggestions from code review

Co-authored-by: Ashutosh Tripathi <39340292+ashucoder9@users.noreply.github.com>

* Apply suggestions from code review

* Update validator-staking-operations.md

* Keep binary and ansible installation

Co-authored-by: Danny Salman <salman.danny03@hotmail.com>
Co-authored-by: Nazeim <nazemfaour89@gmail.com>
Co-authored-by: Ashutosh Tripathi <39340292+ashucoder9@users.noreply.github.com>