Added Further Reading (#1403)
Hey there - I was looking into Polygon's zkEVM and lacked a technical understanding of the rollup and how to deploy it. I found the above links extremely useful and adding them to the page will ease navigation and usability for developers. They strengthen and build upon the content provided in the rest of the page.